+++ cumulusci/tasks/metadata_etl/tests/test_layouts.py
@@ -1,6 +1,6 @@
 from cumulusci.tasks.salesforce.tests.util import create_task
 from cumulusci.tasks.metadata_etl import AddRelatedLists
-from cumulusci.util.xml import metadata_tree
+from cumulusci.utils.xml import metadata_tree
 
 MD = "{%s}" % metadata_tree.METADATA_NAMESPACE
