+++ cumulusci/tasks/metadata_etl/tests/test_sharing.py
@@ -6,7 +6,7 @@ import pytest
 from cumulusci.core.exceptions import TaskOptionsError, CumulusCIException
 from cumulusci.tasks.salesforce.tests.util import create_task
 from cumulusci.tasks.metadata_etl import SetOrgWideDefaults
-from cumulusci.util.xml import metadata_tree
+from cumulusci.utils.xml import metadata_tree
 
 MD = "{%s}" % metadata_tree.METADATA_NAMESPACE
