+++ cumulusci/tasks/metadata_etl/tests/test_value_sets.py
@@ -4,7 +4,7 @@ import pytest
 from cumulusci.core.exceptions import TaskOptionsError
 from cumulusci.tasks.salesforce.tests.util import create_task
 from cumulusci.tasks.metadata_etl import AddValueSetEntries
-from cumulusci.util.xml import metadata_tree
+from cumulusci.utils.xml import metadata_tree
 
 MD = "{%s}" % "http://soap.sforce.com/2006/04/metadata"
 VALUESET_XML = b"""<?xml version="1.0" encoding="UTF-8"?>
